+++ orange/src/components/Necessity/NecessityCreateModal.tsx
@@ -1,59 +1,8 @@
 import React, { Component, Dispatch } from "react";
 import { connect } from "react-redux";
-import styled, { css } from 'styled-components';
 import { necessityActions } from "../../store/actions";
 import { necessityStatus } from "../../constants/constants";
 import "./NecessityCreateModal.css";
-
-interface ICircleButtonProps {
-    transition? : any;
-    open?: any;
-}
-
-const CircleButton = styled.button<ICircleButtonProps>`
-  	background: #38d9a9;
-  	&:hover {
-  	 	 background: #63e6be;
-  	}
-  	&:active {
-   		 background: #20c997;
-  	}
-
-  	z-index: 5;
-  	cursor: pointer;
-  	width: 80px;
- 	height: 80px;
-  	display: block;
-  	align-items: center;
-  	justify-content: center;
-  	font-size: 60px;
-  	position: absolute;
-  	left: 50%;
-  	bottom: 0px;
-  	transform: translate(-50%, 50%);
-	color: white;
-	border-radius: 50%;
-	border: none;
-	outline: none;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-
-	transition: 0.125s all ease-in;
-	${props =>
-		props.open &&
-		css`
-		background: #ff6b6b;
-		&:hover {
-			background: #ff8787;
-		}
-		&:active {
-			background: #fa5252;
-		}
-		transform: translate(-50%, 50%) rotate(45deg);
-    `}
-`;
-
 
 interface Props {
     history: any;
@@ -161,14 +110,14 @@ class NecessityCreateModal extends Component<Props, State> {
                 />
 
                 <div className="clearfix">
-                <CircleButton
+                <button
                     type="button"
                     className="createbtn"
                     disabled={this.state.name==="" || this.state.option==="" || this.state.description==="" || this.state.price===""}
                     onClick={this.create}
                 >
                     생필품 입력
-                </CircleButton>
+                </button>
             </div>
           </div>
         </form>
